requirements.txt
python-dotenv
SQLAlchemy
fastapi
asyncpg
uvicorn
pydantic
python-jose
nh3
jinja2
asyncpg
greenlet
aiobotocore
fastapi_users
fastapi_users_db_sqlalchemy
bs4
setuptools
passlib
bcrypt==4.0.1